altered swiftaccessIO to accept DvObjects
