+++ src/main/java/edu/harvard/iq/dataverse/dataaccess/InputStreamIO.java
@@ -138,9 +138,6 @@ public class InputStreamIO extends DataFileIO {
         throw new UnsupportedDataAccessOperationException("InputStreamIO: there is no output stream associated with this object.");
     }
 
-    @Override
-    public SwiftAccessIO.DvObjectType getDvObjectType() throws IOException {
-        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
-    }
+   
 
 }
